feat: Simplify circle endpoints
